+++ eras/alonzo/test-suite/test/Test/Cardano/Ledger/Alonzo/Serialisation/CDDL.hs
@@ -7,11 +7,10 @@ module Test.Cardano.Ledger.Alonzo.Serialisation.CDDL
   )
 where
 
-import Cardano.Ledger.Alonzo (AlonzoEra)
+import Cardano.Ledger.Alonzo (Alonzo)
 import Cardano.Ledger.Alonzo.Data (Data)
 import Cardano.Ledger.Alonzo.TxWits (AlonzoTxWits, Redeemers)
 import Cardano.Ledger.Core
-import Cardano.Ledger.Crypto (StandardCrypto)
 import qualified Cardano.Ledger.ShelleyMA.Timelocks as MA
 import qualified Data.ByteString.Lazy as BSL
 import Paths_cardano_ledger_alonzo_test
@@ -21,21 +20,19 @@ import Test.Cardano.Ledger.Shelley.Serialisation.CDDLUtils
   )
 import Test.Tasty (TestTree, testGroup, withResource)
 
-type A = AlonzoEra StandardCrypto
-
 tests :: Int -> TestTree
 tests n = withResource combinedCDDL (const (pure ())) $ \cddl ->
   testGroup "CDDL roundtrip tests" $
-    [ cddlTest @(Value A) n "coin",
-      cddlAnnotatorTest @(TxBody A) n "transaction_body",
-      cddlAnnotatorTest @(AuxiliaryData A) n "auxiliary_data",
-      cddlAnnotatorTest @(MA.Timelock (AlonzoEra StandardCrypto)) n "native_script",
-      cddlAnnotatorTest @(Data A) n "plutus_data",
-      cddlTest @(TxOut A) n "transaction_output",
-      cddlAnnotatorTest @(AlonzoTxWits A) n "transaction_witness_set",
-      cddlTest @(PParamsUpdate A) n "protocol_param_update",
-      cddlAnnotatorTest @(Redeemers A) n "[* redeemer]",
-      cddlAnnotatorTest @(Tx A) n "transaction"
+    [ cddlTest @(Value Alonzo) n "coin",
+      cddlAnnotatorTest @(TxBody Alonzo) n "transaction_body",
+      cddlAnnotatorTest @(AuxiliaryData Alonzo) n "auxiliary_data",
+      cddlAnnotatorTest @(MA.Timelock Alonzo) n "native_script",
+      cddlAnnotatorTest @(Data Alonzo) n "plutus_data",
+      cddlTest @(TxOut Alonzo) n "transaction_output",
+      cddlAnnotatorTest @(AlonzoTxWits Alonzo) n "transaction_witness_set",
+      cddlTest @(PParamsUpdate Alonzo) n "protocol_param_update",
+      cddlAnnotatorTest @(Redeemers Alonzo) n "[* redeemer]",
+      cddlAnnotatorTest @(Tx Alonzo) n "transaction"
     ]
       <*> pure cddl