+++ eras/shelley-ma/test-suite/test/Test/Cardano/Ledger/Allegra/ScriptTranslation.hs
@@ -5,9 +5,12 @@ module Test.Cardano.Ledger.Allegra.ScriptTranslation
   )
 where
 
+import Cardano.Ledger.Allegra (Allegra)
 import Cardano.Ledger.Allegra.Translation ()
 import Cardano.Ledger.Block (txid)
 import Cardano.Ledger.Core (TranslateEra (..), hashScript)
+import Cardano.Ledger.Crypto (StandardCrypto)
+import Cardano.Ledger.Shelley (Shelley)
 import qualified Cardano.Ledger.Shelley.API as S
 import Cardano.Ledger.Shelley.LedgerState (LedgerState (..))
 import Cardano.Ledger.Shelley.PParams (emptyPParams)
@@ -20,18 +23,9 @@ import Data.Default.Class (def)
 import qualified Data.Map.Strict as Map
 import qualified Data.Sequence.Strict as StrictSeq
 import qualified Data.Set as Set
-import Test.Cardano.Ledger.EraBuffet
-  ( AllegraEra,
-    ShelleyEra,
-    StandardCrypto,
-  )
 import Test.Cardano.Ledger.Shelley.Utils (applySTSTest, runShelleyBase)
 import Test.Tasty (TestTree)
 import Test.Tasty.HUnit (testCase)
-
-type Allegra = AllegraEra StandardCrypto
-
-type Shelley = ShelleyEra StandardCrypto
 
 bootstrapTxId :: S.TxId StandardCrypto
 bootstrapTxId = txid @Shelley txb
@@ -51,7 +45,7 @@ fromRight :: Either e a -> a
 fromRight (Right x) = x
 fromRight _ = undefined
 
-script :: S.MultiSig (ShelleyEra StandardCrypto)
+script :: S.MultiSig Shelley
 script = S.RequireAllOf []
 
 scriptHash :: S.ScriptHash StandardCrypto