+++ eras/shelley-ma/test-suite/test/Test/Cardano/Ledger/ShelleyMA/Serialisation/CDDL.hs
@@ -6,33 +6,28 @@ module Test.Cardano.Ledger.ShelleyMA.Serialisation.CDDL
   )
 where
 
-import Cardano.Ledger.Allegra (AllegraEra)
+import Cardano.Ledger.Allegra (Allegra)
 import qualified Cardano.Ledger.Core as Core
-import Cardano.Ledger.Crypto (StandardCrypto)
-import Cardano.Ledger.Mary (MaryEra)
+import Cardano.Ledger.Mary (Mary)
 import qualified Data.ByteString.Lazy as BSL
-import Paths_cardano_ledger_shelley_ma_test
+import Paths_cardano_ledger_shelley_ma_test (getDataFileName)
 import Test.Cardano.Ledger.Shelley.Serialisation.CDDLUtils
   ( cddlAnnotatorTest,
     cddlTest,
   )
 import Test.Tasty (TestTree, testGroup, withResource)
 
-type A = AllegraEra StandardCrypto
-
-type M = MaryEra StandardCrypto
-
 cddlTests :: Int -> TestTree
 cddlTests n = withResource combinedCDDL (const (pure ())) $ \cddl ->
   testGroup "CDDL roundtrip tests" $
-    [ cddlTest @(Core.Value A) n "coin",
-      cddlTest @(Core.Value M) n "value",
-      cddlAnnotatorTest @(Core.TxBody M) n "transaction_body",
-      cddlAnnotatorTest @(Core.TxBody A) n "transaction_body_allegra",
-      cddlAnnotatorTest @(Core.Script M) n "native_script",
-      cddlAnnotatorTest @(Core.Script A) n "native_script",
-      cddlAnnotatorTest @(Core.AuxiliaryData M) n "auxiliary_data",
-      cddlAnnotatorTest @(Core.AuxiliaryData A) n "auxiliary_data"
+    [ cddlTest @(Core.Value Allegra) n "coin",
+      cddlTest @(Core.Value Mary) n "value",
+      cddlAnnotatorTest @(Core.TxBody Mary) n "transaction_body",
+      cddlAnnotatorTest @(Core.TxBody Allegra) n "transaction_body_allegra",
+      cddlAnnotatorTest @(Core.Script Mary) n "native_script",
+      cddlAnnotatorTest @(Core.Script Allegra) n "native_script",
+      cddlAnnotatorTest @(Core.AuxiliaryData Mary) n "auxiliary_data",
+      cddlAnnotatorTest @(Core.AuxiliaryData Allegra) n "auxiliary_data"
     ]
       <*> pure cddl