+++ eras/shelley/test-suite/test/Test/Cardano/Ledger/Shelley/Fees.hs
@@ -24,9 +24,7 @@ import Cardano.Ledger.Coin (Coin (..))
 import Cardano.Ledger.Core
 import qualified Cardano.Ledger.Crypto as Cr
 import Cardano.Ledger.Keys
-  ( DSignable,
-    Hash,
-    KeyHash,
+  ( KeyHash,
     KeyPair (..),
     KeyRole (..),
     asWitness,
@@ -34,7 +32,7 @@ import Cardano.Ledger.Keys
     vKey,
   )
 import Cardano.Ledger.SafeHash (hashAnnotated)
-import Cardano.Ledger.Shelley (ShelleyEra)
+import Cardano.Ledger.Shelley (Shelley)
 import Cardano.Ledger.Shelley.API
   ( Addr,
     Credential (..),
@@ -74,12 +72,10 @@ import qualified Data.ByteString.Char8 as BS (pack)
 import qualified Data.ByteString.Lazy as BSL
 import qualified Data.Map.Strict as Map (empty, singleton)
 import Data.Maybe (fromJust)
-import Data.Proxy
 import qualified Data.Sequence.Strict as StrictSeq
 import qualified Data.Set as Set
-import GHC.Stack
+import GHC.Stack (HasCallStack)
 import Lens.Micro
-import Test.Cardano.Ledger.Shelley.ConcreteCryptoTypes (Mock, StandardCrypto)
 import Test.Cardano.Ledger.Shelley.Generator.EraGen (genesisId)
 import Test.Cardano.Ledger.Shelley.Generator.ShelleyEraGen ()
 import Test.Cardano.Ledger.Shelley.Utils
@@ -92,13 +88,8 @@ import Test.Cardano.Ledger.Shelley.Utils
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.HUnit (Assertion, testCase, (@?=))
 
-sizeTest ::
-  (HasCallStack, Cr.Crypto c) =>
-  proxy c ->
-  BSL.ByteString ->
-  ShelleyTx (ShelleyEra c) ->
-  Assertion
-sizeTest _ b16 tx = do
+sizeTest :: HasCallStack => BSL.ByteString -> ShelleyTx Shelley -> Assertion
+sizeTest b16 tx = do
   Base16.encode (serialize tx) @?= b16
   (tx ^. sizeTxF) @?= toInteger (BSL.length b16 `div` 2)
 
@@ -174,7 +165,7 @@ carlPay = KeyPair vk sk
 
 -- | Simple Transaction which consumes one UTxO and creates one UTxO
 -- | and has one witness
-txbSimpleUTxO :: forall c. Cr.Crypto c => ShelleyTxBody (ShelleyEra c)
+txbSimpleUTxO :: ShelleyTxBody Shelley
 txbSimpleUTxO =
   ShelleyTxBody
     { _inputs = Set.fromList [TxIn genesisId minBound],
@@ -187,16 +178,13 @@ txbSimpleUTxO =
       _mdHash = SNothing
     }
 
--- | to use makeWitnessesVKey, we need to know we can sign the TxBody for that (ShelleyEra c)
-type BodySignable era = DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody)
-
-txSimpleUTxO :: forall c. (Cr.Crypto c, BodySignable (ShelleyEra c)) => ShelleyTx (ShelleyEra c)
+txSimpleUTxO :: ShelleyTx Shelley
 txSimpleUTxO =
   ShelleyTx
     { body = txbSimpleUTxO,
       wits =
         mempty
-          { addrWits = makeWitnessesVKey (hashAnnotated $ txbSimpleUTxO @c) [alicePay]
+          { addrWits = makeWitnessesVKey (hashAnnotated txbSimpleUTxO) [alicePay]
           },
       auxiliaryData = SNothing
     }
@@ -206,7 +194,7 @@ txSimpleUTxOBytes16 = "83a4008182582003170a2e7597b7b7e3d84c05391d139a62b157e7878
 
 -- | Transaction which consumes two UTxO and creates five UTxO
 -- | and has two witness
-txbMutiUTxO :: forall c. Cr.Crypto c => ShelleyTxBody (ShelleyEra c)
+txbMutiUTxO :: ShelleyTxBody Shelley
 txbMutiUTxO =
   ShelleyTxBody
     { _inputs =
@@ -230,7 +218,7 @@ txbMutiUTxO =
       _mdHash = SNothing
     }
 
-txMutiUTxO :: forall c. (Cr.Crypto c, BodySignable (ShelleyEra c)) => ShelleyTx (ShelleyEra c)
+txMutiUTxO :: ShelleyTx Shelley
 txMutiUTxO =
   ShelleyTx
     { body = txbMutiUTxO,
@@ -238,7 +226,7 @@ txMutiUTxO =
         mempty
           { addrWits =
               makeWitnessesVKey
-                (hashAnnotated $ txbMutiUTxO @c)
+                (hashAnnotated txbMutiUTxO)
                 [ alicePay,
                   bobPay
                 ]
@@ -250,7 +238,7 @@ txMutiUTxOBytes16 :: BSL.ByteString
 txMutiUTxOBytes16 = "83a4008282582003170a2e7597b7b7e3d84c05391d139a62b157e78786d8c082f29dcf4c1113140082582003170a2e7597b7b7e3d84c05391d139a62b157e78786d8c082f29dcf4c11131401018582583900e9686d801fa32aeb4390c2f2a53bb0314a9c744c46a2cada394a371fc6852b6aaed73bcf346a57ef99adae3000b51c7c59faaeb15993df760a82583900e9686d801fa32aeb4390c2f2a53bb0314a9c744c46a2cada394a371fc6852b6aaed73bcf346a57ef99adae3000b51c7c59faaeb15993df761482583900e9686d801fa32aeb4390c2f2a53bb0314a9c744c46a2cada394a371fc6852b6aaed73bcf346a57ef99adae3000b51c7c59faaeb15993df76181e825839000d2a471489a90f2910ec67ded8e215bfcd669bae77e7f9ab15850abd4e130c0bdeb7768edf2e8f85007fd52073e3dc1871f4c47f9dfca92e1828825839000d2a471489a90f2910ec67ded8e215bfcd669bae77e7f9ab15850abd4e130c0bdeb7768edf2e8f85007fd52073e3dc1871f4c47f9dfca92e18320218c7030aa1008282582037139648f2c22bbf1d0ef9af37cfebc9014b1e0e2a55be87c4b3b231a8d84d2658405ef09b22172cd28678e76e600e899886852e03567e2e72b4815629471e736a0cd424dc71cdaa0d0403371d79ea3d0cb7f28cb0740ebfcd8947343eba99a6aa088258204628aaf16d6e1baa061d1296419542cb09287c639163d0fdbdac0ff23699797e5840ea98ef8052776aa5c182621cfd2ec91011d327527fc2531be9e1a8356c10f25f3fe5a5a7f549a0dc3b17c4ad8e4b8673b63a87977ac899b675f3ce3d6badae01f6"
 
 -- | Transaction which registers a stake key
-txbRegisterStake :: forall c. Cr.Crypto c => ShelleyTxBody (ShelleyEra c)
+txbRegisterStake :: ShelleyTxBody Shelley
 txbRegisterStake =
   ShelleyTxBody
     { _inputs = Set.fromList [TxIn genesisId minBound],
@@ -263,13 +251,13 @@ txbRegisterStake =
       _mdHash = SNothing
     }
 
-txRegisterStake :: forall c. (Cr.Crypto c, BodySignable (ShelleyEra c)) => ShelleyTx (ShelleyEra c)
+txRegisterStake :: ShelleyTx Shelley
 txRegisterStake =
   ShelleyTx
     { body = txbRegisterStake,
       wits =
         mempty
-          { addrWits = makeWitnessesVKey (hashAnnotated $ txbRegisterStake @c) [alicePay]
+          { addrWits = makeWitnessesVKey (hashAnnotated txbRegisterStake) [alicePay]
           },
       auxiliaryData = SNothing
     }
@@ -278,7 +266,7 @@ txRegisterStakeBytes16 :: BSL.ByteString
 txRegisterStakeBytes16 = "83a5008182582003170a2e7597b7b7e3d84c05391d139a62b157e78786d8c082f29dcf4c11131400018182583900e9686d801fa32aeb4390c2f2a53bb0314a9c744c46a2cada394a371fc6852b6aaed73bcf346a57ef99adae3000b51c7c59faaeb15993df760a02185e030a048182008200581cc6852b6aaed73bcf346a57ef99adae3000b51c7c59faaeb15993df76a100818258204628aaf16d6e1baa061d1296419542cb09287c639163d0fdbdac0ff23699797e58403271792b002eb39bcb133668e851a5ffba9c13ad2b5c5a7bbc850a17de8309cbb9649d9e90eb4c9cc82f28f204408d513ccc575ce1f61808f67793429ff1880ef6"
 
 -- | Transaction which delegates a stake key
-txbDelegateStake :: forall c. Cr.Crypto c => ShelleyTxBody (ShelleyEra c)
+txbDelegateStake :: ShelleyTxBody Shelley
 txbDelegateStake =
   ShelleyTxBody
     { _inputs = Set.fromList [TxIn genesisId minBound],
@@ -295,7 +283,7 @@ txbDelegateStake =
       _mdHash = SNothing
     }
 
-txDelegateStake :: forall c. (Cr.Crypto c, BodySignable (ShelleyEra c)) => ShelleyTx (ShelleyEra c)
+txDelegateStake :: ShelleyTx Shelley
 txDelegateStake =
   ShelleyTx
     { body = txbDelegateStake,
@@ -303,7 +291,7 @@ txDelegateStake =
         mempty
           { addrWits =
               makeWitnessesVKey
-                (hashAnnotated $ txbDelegateStake @c)
+                (hashAnnotated txbDelegateStake)
                 [asWitness alicePay, asWitness bobStake]
           },
       auxiliaryData = SNothing
@@ -313,7 +301,7 @@ txDelegateStakeBytes16 :: BSL.ByteString
 txDelegateStakeBytes16 = "83a5008182582003170a2e7597b7b7e3d84c05391d139a62b157e78786d8c082f29dcf4c11131400018182583900e9686d801fa32aeb4390c2f2a53bb0314a9c744c46a2cada394a371fc6852b6aaed73bcf346a57ef99adae3000b51c7c59faaeb15993df760a02185e030a048183028200581c4e130c0bdeb7768edf2e8f85007fd52073e3dc1871f4c47f9dfca92e581c5d43e1f1048b2619f51abc0cf505e4d4f9cb84becefd468d1a2fe335a100828258209921fa37a7d167aab519bb937d7ac6e522ad6d259a6173523357b971e05f41ff58403bad563c201b4f62448db12711af2d916776194b5176e9d312d07a328ce7780a63032dce887abc67985629b7aeabb0c334e84094f44d7e51ae51b5c799a83c0d8258204628aaf16d6e1baa061d1296419542cb09287c639163d0fdbdac0ff23699797e584064aef85b046d2d0072cd64844e9f13d86651a1db74d356a10ecd7fb35a664fc466e543ea55cfbffd74025dc092d62c4b22d7e2de4decb4f049df354cfae9790af6"
 
 -- | Transaction which de-registers a stake key
-txbDeregisterStake :: forall c. Cr.Crypto c => ShelleyTxBody (ShelleyEra c)
+txbDeregisterStake :: ShelleyTxBody Shelley
 txbDeregisterStake =
   ShelleyTxBody
     { _inputs = Set.fromList [TxIn genesisId minBound],
@@ -326,7 +314,7 @@ txbDeregisterStake =
       _mdHash = SNothing
     }
 
-txDeregisterStake :: forall c. (Cr.Crypto c, BodySignable (ShelleyEra c)) => ShelleyTx (ShelleyEra c)
+txDeregisterStake :: ShelleyTx Shelley
 txDeregisterStake =
   ShelleyTx
     { body = txbDeregisterStake,
@@ -334,8 +322,8 @@ txDeregisterStake =
         mempty
           { addrWits =
               makeWitnessesVKey
-                (hashAnnotated $ txbDeregisterStake @c)
-                [alicePay @(EraCrypto (ShelleyEra c))]
+                (hashAnnotated txbDeregisterStake)
+                [alicePay @(EraCrypto Shelley)]
           },
       auxiliaryData = SNothing
     }
@@ -344,7 +332,7 @@ txDeregisterStakeBytes16 :: BSL.ByteString
 txDeregisterStakeBytes16 = "83a5008182582003170a2e7597b7b7e3d84c05391d139a62b157e78786d8c082f29dcf4c11131400018182583900e9686d801fa32aeb4390c2f2a53bb0314a9c744c46a2cada394a371fc6852b6aaed73bcf346a57ef99adae3000b51c7c59faaeb15993df760a02185e030a048182018200581cc6852b6aaed73bcf346a57ef99adae3000b51c7c59faaeb15993df76a100818258204628aaf16d6e1baa061d1296419542cb09287c639163d0fdbdac0ff23699797e5840409db925fa592b7f4c76e44d738789f4b0ffb2b9cf4567af127121d635491b4eb736e8c92571f1329f14d06aad7ec42ca654ae65eb63b0b01d30cc4454aee80cf6"
 
 -- | Transaction which registers a stake pool
-txbRegisterPool :: Cr.Crypto c => ShelleyTxBody (ShelleyEra c)
+txbRegisterPool :: ShelleyTxBody Shelley
 txbRegisterPool =
   ShelleyTxBody
     { _inputs = Set.fromList [TxIn genesisId minBound],
@@ -357,13 +345,13 @@ txbRegisterPool =
       _mdHash = SNothing
     }
 
-txRegisterPool :: forall c. (Cr.Crypto c, BodySignable (ShelleyEra c)) => ShelleyTx (ShelleyEra c)
+txRegisterPool :: ShelleyTx Shelley
 txRegisterPool =
   ShelleyTx
     { body = txbRegisterPool,
       wits =
         mempty
-          { addrWits = makeWitnessesVKey (hashAnnotated $ txbRegisterPool @c) [alicePay]
+          { addrWits = makeWitnessesVKey (hashAnnotated txbRegisterPool) [alicePay]
           },
       auxiliaryData = SNothing
     }
@@ -372,7 +360,7 @@ txRegisterPoolBytes16 :: BSL.ByteString
 txRegisterPoolBytes16 = "83a5008182582003170a2e7597b7b7e3d84c05391d139a62b157e78786d8c082f29dcf4c11131400018182583900e9686d801fa32aeb4390c2f2a53bb0314a9c744c46a2cada394a371fc6852b6aaed73bcf346a57ef99adae3000b51c7c59faaeb15993df760a02185e030a04818a03581c5d43e1f1048b2619f51abc0cf505e4d4f9cb84becefd468d1a2fe33558208e61e1fa4855ea3aa0b8881a9e2e453c8c73536bdaabb64d36de86ee5a02519a0105d81e82010a581de0c6852b6aaed73bcf346a57ef99adae3000b51c7c59faaeb15993df7681581cc6852b6aaed73bcf346a57ef99adae3000b51c7c59faaeb15993df76818301f66872656c61792e696f826a616c6963652e706f6f6c427b7da100818258204628aaf16d6e1baa061d1296419542cb09287c639163d0fdbdac0ff23699797e5840165c6aa107571daafb1f9093d3cdc184a4068e8ff9243715c13335feb3652dc0d817b3b015a9929c9d83a0dd406fe71658fdccbf7925d2fff316237b499c2003f6"
 
 -- | Transaction which retires a stake pool
-txbRetirePool :: forall c. Cr.Crypto c => ShelleyTxBody (ShelleyEra c)
+txbRetirePool :: ShelleyTxBody Shelley
 txbRetirePool =
   ShelleyTxBody
     { _inputs = Set.fromList [TxIn genesisId minBound],
@@ -385,13 +373,13 @@ txbRetirePool =
       _mdHash = SNothing
     }
 
-txRetirePool :: forall c. (Cr.Crypto c, BodySignable (ShelleyEra c)) => ShelleyTx (ShelleyEra c)
+txRetirePool :: ShelleyTx Shelley
 txRetirePool =
   ShelleyTx
     { body = txbRetirePool,
       wits =
         mempty
-          { addrWits = makeWitnessesVKey (hashAnnotated $ txbRetirePool @c) [alicePay]
+          { addrWits = makeWitnessesVKey (hashAnnotated txbRetirePool) [alicePay]
           },
       auxiliaryData = SNothing
     }
@@ -404,7 +392,7 @@ txRetirePoolBytes16 = "83a5008182582003170a2e7597b7b7e3d84c05391d139a62b157e7878
 md :: MD.Metadata era
 md = MD.Metadata $ Map.singleton 0 (MD.List [MD.I 5, MD.S "hello"])
 
-txbWithMD :: forall c. Cr.Crypto c => ShelleyTxBody (ShelleyEra c)
+txbWithMD :: ShelleyTxBody Shelley
 txbWithMD =
   ShelleyTxBody
     { _inputs = Set.fromList [TxIn genesisId minBound],
@@ -414,16 +402,16 @@ txbWithMD =
       _txfee = Coin 94,
       _ttl = SlotNo 10,
       _txUpdate = SNothing,
-      _mdHash = SJust $ hashAuxiliaryData @(ShelleyEra c) md
+      _mdHash = SJust $ hashAuxiliaryData @Shelley md
     }
 
-txWithMD :: forall c. (Cr.Crypto c, BodySignable (ShelleyEra c)) => ShelleyTx (ShelleyEra c)
+txWithMD :: ShelleyTx Shelley
 txWithMD =
   ShelleyTx
     { body = txbWithMD,
       wits =
         mempty
-          { addrWits = makeWitnessesVKey (hashAnnotated $ txbWithMD @c) [alicePay]
+          { addrWits = makeWitnessesVKey (hashAnnotated txbWithMD) [alicePay]
           },
       auxiliaryData = SJust md
     }
@@ -441,7 +429,7 @@ msig =
       (RequireSignature . asWitness . hashKey . vKey) carlPay
     ]
 
-txbWithMultiSig :: Cr.Crypto c => ShelleyTxBody (ShelleyEra c)
+txbWithMultiSig :: ShelleyTxBody Shelley
 txbWithMultiSig =
   ShelleyTxBody
     { _inputs = Set.fromList [TxIn genesisId minBound], -- acting as if this is multi-sig
@@ -454,7 +442,7 @@ txbWithMultiSig =
       _mdHash = SNothing
     }
 
-txWithMultiSig :: forall c. Mock c => ShelleyTx (ShelleyEra c)
+txWithMultiSig :: ShelleyTx Shelley
 txWithMultiSig =
   ShelleyTx
     { body = txbWithMultiSig,
@@ -462,9 +450,9 @@ txWithMultiSig =
         mempty
           { addrWits =
               makeWitnessesVKey
-                (hashAnnotated $ txbWithMultiSig @c)
+                (hashAnnotated txbWithMultiSig)
                 [alicePay, bobPay],
-            scriptWits = Map.singleton (hashScript @(ShelleyEra c) msig) msig
+            scriptWits = Map.singleton (hashScript @Shelley msig) msig
           },
       auxiliaryData = SNothing
     }
@@ -473,7 +461,7 @@ txWithMultiSigBytes16 :: BSL.ByteString
 txWithMultiSigBytes16 = "83a4008182582003170a2e7597b7b7e3d84c05391d139a62b157e78786d8c082f29dcf4c11131400018182583900e9686d801fa32aeb4390c2f2a53bb0314a9c744c46a2cada394a371fc6852b6aaed73bcf346a57ef99adae3000b51c7c59faaeb15993df760a02185e030aa2008282582037139648f2c22bbf1d0ef9af37cfebc9014b1e0e2a55be87c4b3b231a8d84d265840e3b8f50632325fbd1f82202ce5a8b4672bd96c50a338d70c0aa96720f6f7fbf60e0ce708f3a7e28faa0d78dc437a0b61e02205ddb1db22d02ba35b37a7fe03068258204628aaf16d6e1baa061d1296419542cb09287c639163d0fdbdac0ff23699797e584089c20cb6246483bbd0b2006f658597eff3e8ab3b8a6e9b22cb3c5b95cf0d3a2b96107acef88319fa2dd0fb28adcfdb330bb99f1f0058918a75d951ca9b73660c0181830302838200581ce9686d801fa32aeb4390c2f2a53bb0314a9c744c46a2cada394a371f8200581c0d2a471489a90f2910ec67ded8e215bfcd669bae77e7f9ab15850abd8200581cd0671052191a58c554eee27808b2b836a03ca369ca7a847f8c37d6f9f6"
 
 -- | Transaction with a Reward Withdrawal
-txbWithWithdrawal :: Cr.Crypto c => ShelleyTxBody (ShelleyEra c)
+txbWithWithdrawal :: ShelleyTxBody Shelley
 txbWithWithdrawal =
   ShelleyTxBody
     { _inputs = Set.fromList [TxIn genesisId minBound],
@@ -486,7 +474,7 @@ txbWithWithdrawal =
       _mdHash = SNothing
     }
 
-txWithWithdrawal :: forall c. (Cr.Crypto c, BodySignable (ShelleyEra c)) => ShelleyTx (ShelleyEra c)
+txWithWithdrawal :: ShelleyTx Shelley
 txWithWithdrawal =
   ShelleyTx
     { body = txbWithWithdrawal,
@@ -494,7 +482,7 @@ txWithWithdrawal =
         mempty
           { addrWits =
               makeWitnessesVKey
-                (hashAnnotated $ txbWithWithdrawal @c)
+                (hashAnnotated txbWithWithdrawal)
                 [asWitness alicePay, asWitness aliceStake]
           },
       auxiliaryData = SNothing
@@ -507,11 +495,11 @@ txWithWithdrawalBytes16 = "83a5008182582003170a2e7597b7b7e3d84c05391d139a62b157e
 -- given minfeeA and minfeeB are set to 1.
 testEvaluateTransactionFee :: Assertion
 testEvaluateTransactionFee =
-  API.evaluateTransactionFee @(ShelleyEra StandardCrypto)
+  API.evaluateTransactionFee @Shelley
     pp
     txSimpleUTxONoWit
     1
-    @?= getMinFeeTx pp (txSimpleUTxO @StandardCrypto)
+    @?= getMinFeeTx pp txSimpleUTxO
   where
     pp = emptyPParams {_minfeeA = 1, _minfeeB = 1}
 
@@ -533,18 +521,15 @@ sizeTests :: TestTree
 sizeTests =
   testGroup
     "Fee Tests"
-    [ testCase "simple utxo" $ sizeTest p txSimpleUTxOBytes16 txSimpleUTxO,
-      testCase "multiple utxo" $ sizeTest p txMutiUTxOBytes16 txMutiUTxO,
-      testCase "register stake key" $ sizeTest p txRegisterStakeBytes16 txRegisterStake,
-      testCase "delegate stake key" $ sizeTest p txDelegateStakeBytes16 txDelegateStake,
-      testCase "deregister stake key" $ sizeTest p txDeregisterStakeBytes16 txDeregisterStake,
-      testCase "register stake pool" $ sizeTest p txRegisterPoolBytes16 txRegisterPool,
-      testCase "retire stake pool" $ sizeTest p txRetirePoolBytes16 txRetirePool,
-      testCase "auxiliaryData" $ sizeTest p txWithMDBytes16 txWithMD,
-      testCase "multisig" $ sizeTest p txWithMultiSigBytes16 txWithMultiSig,
-      testCase "reward withdrawal" $ sizeTest p txWithWithdrawalBytes16 txWithWithdrawal,
+    [ testCase "simple utxo" $ sizeTest txSimpleUTxOBytes16 txSimpleUTxO,
+      testCase "multiple utxo" $ sizeTest txMutiUTxOBytes16 txMutiUTxO,
+      testCase "register stake key" $ sizeTest txRegisterStakeBytes16 txRegisterStake,
+      testCase "delegate stake key" $ sizeTest txDelegateStakeBytes16 txDelegateStake,
+      testCase "deregister stake key" $ sizeTest txDeregisterStakeBytes16 txDeregisterStake,
+      testCase "register stake pool" $ sizeTest txRegisterPoolBytes16 txRegisterPool,
+      testCase "retire stake pool" $ sizeTest txRetirePoolBytes16 txRetirePool,
+      testCase "auxiliaryData" $ sizeTest txWithMDBytes16 txWithMD,
+      testCase "multisig" $ sizeTest txWithMultiSigBytes16 txWithMultiSig,
+      testCase "reward withdrawal" $ sizeTest txWithWithdrawalBytes16 txWithWithdrawal,
       testCase "evaluate transaction fee" testEvaluateTransactionFee
     ]
-  where
-    p :: Proxy StandardCrypto
-    p = Proxy