+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Era.hs
@@ -2,28 +2,39 @@ module Cardano.Ledger.Api.Era
   ( -- * Eras
 
     -- ** Shelley
+    Shelley,
     ShelleyEra,
 
-    -- ** Shelley MA
-    ShelleyMAEra,
-    MaryEra,
+    -- ** Allegra
+    Allegra,
     AllegraEra,
 
+    -- ** Mary
+    Mary,
+    MaryEra,
+
     -- ** Alonzo
+    Alonzo,
     AlonzoEra,
 
     -- ** Babbage
+    Babbage,
     BabbageEra,
 
     -- ** Conway
+    Conway,
     ConwayEra,
+
+    -- * Crypto
+    StandardCrypto,
+    Crypto (..),
   )
 where
 
-import Cardano.Ledger.Allegra (AllegraEra)
-import Cardano.Ledger.Alonzo (AlonzoEra)
-import Cardano.Ledger.Babbage (BabbageEra)
-import Cardano.Ledger.Conway (ConwayEra)
-import Cardano.Ledger.Mary (MaryEra)
-import Cardano.Ledger.Shelley (ShelleyEra)
-import Cardano.Ledger.ShelleyMA (ShelleyMAEra)
+import Cardano.Ledger.Allegra (Allegra, AllegraEra)
+import Cardano.Ledger.Alonzo (Alonzo, AlonzoEra)
+import Cardano.Ledger.Babbage (Babbage, BabbageEra)
+import Cardano.Ledger.Conway (Conway, ConwayEra)
+import Cardano.Ledger.Crypto (Crypto (..), StandardCrypto)
+import Cardano.Ledger.Mary (Mary, MaryEra)
+import Cardano.Ledger.Shelley (Shelley, ShelleyEra)